update basic linux cmd
