--- lxTQ_ls/NguyenDinhHoangPhuc_19522036.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30204.135
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NguyenDinhHoangPhuc_19522036", "NguyenDinhHoangPhuc_19522036\NguyenDinhHoangPhuc_19522036.csproj", "{16BF3B1C-08E6-4BF0-A922-E7232F3C7C96}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{16BF3B1C-08E6-4BF0-A922-E7232F3C7C96}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{16BF3B1C-08E6-4BF0-A922-E7232F3C7C96}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{16BF3B1C-08E6-4BF0-A922-E7232F3C7C96}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{16BF3B1C-08E6-4BF0-A922-E7232F3C7C96}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0DEB0C37-FAA3-4EBA-A2CC-E5DD0D649F3F}
-	EndGlobalSection
-EndGlobal
