--- lxTQ_ls/NguyenDinhHoangPhuc_19522036/bin/Debug/NguyenDinhHoangPhuc_19522036.exe.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2" />
-    </startup>
-</configuration>
